+++ packages/forms/src/index.ts
@@ -14,3 +14,4 @@
 
 export { AddressForm } from "./address-form";
 export { FormDevTools } from "./form-dev-tools";
+export { OTPForm } from "./otp-form";
